# paddlelabel/openapi.yml
openapi: 3.0.0
info:
  title: PaddleLabel API Specs
  version: 1.0.2
  description: Web backend APIs for PaddleLabel
  contact:
    name: PaddleLabel Developers
    url: 'https://github.com/PaddleCV-SIG/PaddleLabel/issues'
    email: me@linhan.email
  license:
    name: Apache License 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
servers:
  - url: /api
    description: Same origin
paths:
  /projects:
    parameters:
      - $ref: '#/components/parameters/request_id'
    get:
      tags:
        - Project
      summary: 'Read all projects, sort by last modify date'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
      operationId: getAll
      parameters:
        - schema:
            type: string
          in: query
          name: order_by
    post:
      tags:
        - Project
      summary: Create a new project
      description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
      operationId: create
...
  '/projects/{project_id}':
    parameters:
      - $ref: '#/components/parameters/project_id'
      - $ref: '#/components/parameters/request_id'
    get:
      tags:
        - Project
      summary: Get info of a specific project
      description: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not fond
      operationId: get
    delete:
      summary: Delete a project record and ALL TASKS RECORDS under the project.
      description: Delete a project record and ALL TASKS RECORDS under the project. Won't delete file on file system
      tags:
        - Project
      responses:
        '200':
          description: OK
        '404':
          description: No project with such project id
      operationId: remove
    put:
      tags:
        - Project
      summary: Edit multiple project properties
      description: Edit multiple project properties. Empty strings will be set. Properties not in request body won't be changed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: 'Project with project id not fond, or project dont have requested property'
      operationId: update
...
  '/projects/{project_id}/predict':
    parameters:
      - $ref: '#/components/parameters/project_id'
    post:
      summary: Run prediction on all data in the dataset
      tags:
        - Project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                ml_backend_url:
                  type: string
                model:
                  type: string
                same_server:
                  type: boolean
                  description: |-
                    Default false. Will b64 encode image and send to ml backend in this case. Images are directly encoded as file.
                    If the ml backend has read access to project data dir, set same_server to true. This case web backend will send file path to ml backend. Reduce b64 en/decode overhead.
                create_label:
                  type: boolean
                  description: Default false. Create label if model prediction not in current labels. Won't save prediction result in db if prediciont label not present in current labels
              required:
                - ml_backend_url
                - model
      responses:
        '200':
          description: success
      operationId: predict
      description: ''
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: paddlelabel.api.decode_token
  schemas:
    Project:
      type: object
      x-examples:
        Example 1:
          created: '2023-02-09T17:12:26.576325'
          data_dir: /home/lin/3rd_party/3rd_party/labelme/point
          description: null
          labels:
            - color: '#C2939E'
              comment: null
              created: '2023-02-09T17:12:26.602183'
              id: 1
              label_id: 1
              modified: '2023-02-09T17:12:26.602186'
              name: ear
              project_id: 1
              super_category_id: null
            - color: '#FAD858'
              comment: null
              created: '2023-02-09T17:12:26.607238'
              id: 2
              label_id: 2
              modified: '2023-02-09T17:12:26.607243'
              name: hand
              project_id: 1
              super_category_id: null
            - color: '#3BC3F4'
              comment: null
              created: '2023-02-09T17:12:26.613080'
              id: 3
              label_id: 3
              modified: '2023-02-09T17:12:26.613086'
              name: foot
              project_id: 1
              super_category_id: null
          modified: '2023-02-09T17:12:26.576336'
          name: Points - labelme format
          other_settings: null
          project_id: 1
          task_category:
            created: '2023-02-09T07:46:12.059813'
            handler: paddlelabel.task.Point
            modified: '2023-02-09T07:46:12.059818'
            name: point
            task_category_id: 8
          task_category_id: 8
          upid: 952ecbc1a5fc73a3dc7e
      additionalProperties: false
      description: project info and settings
      properties:
        project_id:
          type: integer
          minimum: 1
          format: int32
          example: 3
          description: project表的主键，从1开始
          readOnly: true
        name:
          type: string
          description: 项目名字
          maxLength: 128
          example: A classification project
        description:
          type: string
          description: Optional project description
          maxLength: 256
          example: This project is very cool
        data_dir:
          type: string
          description: Absolute directory path where all the data file is stored
        task_category_id:
          type: integer
          description: 'Top level annotation task category, see TODO for int <-> category map'
        task_category:
          $ref: '#/components/schemas/TaskCategory'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        created:
          type: string
          description: Project creation timestamp in UTC
          format: timestamp
          nullable: true
          readOnly: true
        modified:
          type: string
          description: Last time the project detail or ANY TASK of the project is modified
          format: timestamp
          nullable: true
          readOnly: true
        upid:
          type: string
          description: A string that is unique to each project
        other_settings:
          type: object
          properties:
            mlBackendUrl:
              type: string
              description: 'ML后端地址，一般是 http://localhost:1234'
              example: 'http://localhost:1234'
            perviousModel:
              type: string
              deprecated: true
            models:
              type: object
              deprecated: true
            modelName:
              type: string
              description: 正在使用的模型名称
              example: PicoDet
            labelMapping:
              type: array
              description: 标签对应关系
              items:
                type: object
                properties:
                  model:
                    type: string
                    example: a coco category
                  project:
                    type: string
                    example: a user defined category
            modelSettings:
              type: object
              additionalProperties:
                type: object
                properties:
                  modelFilePath:
                    type: string
                  paramFilePath:
                    type: string
            modelFilePath:
              type: string
            paramFilePath:
              type: string
            lang:
              type: string
            isSample:
              type: boolean
            clasSubCatg:
              type: string
        all_options:
          type: object
          description: This property is only used for passing project create/export options around more easilly. It's discarded after the project is created or exported.
...
    Label:
      title: Label
      type: object
      description: ''
      properties:
        label_id:
          type: integer
        project_id:
          type: integer
        name:
          type: string
        color:
          type: string
        comment:
          type: string
        created:
          type: string
          readOnly: true
        modified:
          type: string
          readOnly: true
        super_category_id:
          type: integer
      required:
        - name
...
    TaskCategory:
      title: TaskCategory
      x-stoplight:
        id: dcukxtsw0qjz2
      type: object
      properties:
        name:
          type: string
        task_category_id:
          type: integer
          format: int32
        handler:
          type: string
          deprecated: true
        created:
          type: string
        modified:
          type: string
      x-examples:
        Example 1:
          name: point
          task_category_id: 8
          handler: paddlelabel.task.Point
          created: '2023-02-09T07:46:12.059813'
          modified: '2023-02-09T07:46:12.059818'
  parameters:
    request_id:
      name: request_id
      in: header
      required: false
      schema:
        type: integer
        maxLength: 30
      description: 'Assign a unique random string each request, backend will reject requests with same request_id within several seconds. Prevent critical operations (mostly post) from executing multiple times'
    project_id:
      name: project_id
      in: path
      schema:
        type: integer
      description: Id of a project
      required: true
...